requirements_stable.txt
streamlit==1.49.1
python-dotenv==1.0.0
langchain==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.0
langchain-community==0.2.0
openai==1.30.0
pinecone-client==3.2.2